+++ mail/mail-common/tests/acceptance/mailbox_conversation.rs
@@ -1,6 +1,8 @@
 use proton_core_api::services::proton::Label as ApiLabel;
 use proton_core_api::services::proton::{LabelId, LabelType as ApiLabelType};
+use proton_core_common::models::ModelExtension;
 use proton_mail_api::services::proton::common::MessageId;
+use proton_mail_api::services::proton::response_data::ConversationLabel as ApiConversationLabel;
 use proton_mail_api::services::proton::response_data::MessageMetadata as ApiMessageMetadata;
 use proton_mail_common::Mailbox;
 use proton_mail_common::datatypes::{ContextualConversation, SystemLabelId};
@@ -107,6 +109,175 @@ async fn test_new_mailbox_sync_conversations() {
     .await
     .unwrap()
     .unwrap();
+}
+
+#[tokio::test]
+async fn test_new_mailbox_syncs_new_conversation_messages_on_push_notification() {
+    // Set up a user and initialise the inbox
+    let ctx = MailTestContext::new().await;
+    let mut params = TestParams::default_basic();
+    let new_label_id = LabelId::from("NEW_LABEL");
+    {
+        let labels = params.labels.get_mut(&ApiLabelType::Label).unwrap();
+
+        labels.push(ApiLabel {
+            id: LabelId::from("testlabel"),
+            name: "testlabel".to_owned(),
+            label_type: ApiLabelType::Label,
+            ..ApiLabel::test_default()
+        });
+
+        labels.push(ApiLabel {
+            id: new_label_id.clone(),
+            name: "testlabel2".to_owned(),
+            label_type: ApiLabelType::Label,
+            ..ApiLabel::test_default()
+        });
+    }
+
+    let message_id1 = MessageId::from("m1");
+    let message_id2 = MessageId::from("m2");
+    let message_id3 = MessageId::from("m3");
+
+    let messages = vec![
+        ApiMessageMetadata {
+            id: message_id1.clone(),
+            conversation_id: params.conversations[0].id.clone(),
+            order: 0,
+            address_id: params.addresses[0].id.clone(),
+            label_ids: vec![LabelId::inbox()],
+            ..ApiMessageMetadata::test_default()
+        },
+        ApiMessageMetadata {
+            id: message_id2.clone(),
+            conversation_id: params.conversations[0].id.clone(),
+            order: 1,
+            address_id: params.addresses[0].id.clone(),
+            label_ids: vec![LabelId::inbox()],
+            ..ApiMessageMetadata::test_default()
+        },
+    ];
+
+    let messages_updated = vec![
+        ApiMessageMetadata {
+            id: message_id1.clone(),
+            conversation_id: params.conversations[0].id.clone(),
+            order: 0,
+            address_id: params.addresses[0].id.clone(),
+            label_ids: vec![LabelId::inbox()],
+            ..ApiMessageMetadata::test_default()
+        },
+        ApiMessageMetadata {
+            id: message_id2.clone(),
+            conversation_id: params.conversations[0].id.clone(),
+            order: 1,
+            address_id: params.addresses[0].id.clone(),
+            label_ids: vec![LabelId::inbox()],
+            ..ApiMessageMetadata::test_default()
+        },
+        ApiMessageMetadata {
+            id: message_id3.clone(),
+            conversation_id: params.conversations[0].id.clone(),
+            order: 2,
+            address_id: params.addresses[0].id.clone(),
+            label_ids: vec![LabelId::inbox()],
+            ..ApiMessageMetadata::test_default()
+        },
+    ];
+
+    let mut conv_updated = params.conversations[0].clone();
+    conv_updated.labels.push(ApiConversationLabel {
+        id: new_label_id.clone(),
+        context_expiration_time: 0,
+        context_num_attachments: 0,
+        context_num_messages: 20,
+        context_num_unread: 0,
+        context_size: 0,
+        context_snooze_time: 0,
+        context_time: 0,
+    });
+
+    let conversations = params.conversations.clone();
+    ctx.setup_user(params.clone()).await;
+    ctx.mock_get_conversations(conversations, 1_u64).await;
+    ctx.mock_get_conversation_messages(params.conversations[0].clone(), messages, 1_u64)
+        .await;
+    let user_ctx = ctx.mail_user_context().await;
+
+    // Create a mailbox
+    let mailbox = Mailbox::with_remote_id(
+        &user_ctx.user_stash().connection().await.unwrap(),
+        LabelId::inbox(),
+    )
+    .await
+    .unwrap();
+
+    // Sync mailbox 1 - this should fire a network request
+    mailbox
+        .sync(
+            &mut user_ctx.user_stash().connection().await.unwrap(),
+            user_ctx.session(),
+            10,
+        )
+        .await
+        .unwrap();
+    let tether = user_ctx.user_stash().connection().await.unwrap();
+    // Get conversations for mailbox.
+    let conversation = Conversation::find_first("", vec![], &tether)
+        .await
+        .unwrap()
+        .unwrap();
+
+    // Get the message for a conversation.
+
+    let result = ContextualConversation::conversation_and_messages(
+        user_ctx.network_monitor_service(),
+        conversation.id(),
+        mailbox.label_id(),
+        user_ctx.user_stash(),
+        user_ctx.session(),
+    )
+    .await
+    .unwrap()
+    .unwrap();
+
+    assert_eq!(result.messages.len(), 2);
+    assert_eq!(result.messages[0].remote_id.as_ref(), Some(&message_id1));
+    assert_eq!(result.messages[1].remote_id.as_ref(), Some(&message_id2));
+
+    ctx.mock_server().reset().await;
+    ctx.mock_get_conversation_messages(conv_updated, messages_updated, 1_u64)
+        .await;
+    ctx.catch_all().await;
+    // Get messages again, should have new message
+    let result = ContextualConversation::conversation_and_messages_from_push_notification(
+        user_ctx.network_monitor_service(),
+        conversation.id(),
+        mailbox.label_id(),
+        user_ctx.user_stash(),
+        user_ctx.session(),
+    )
+    .await
+    .unwrap()
+    .unwrap();
+
+    assert_eq!(result.messages.len(), 3);
+    assert_eq!(result.messages[0].remote_id.as_ref(), Some(&message_id1));
+    assert_eq!(result.messages[1].remote_id.as_ref(), Some(&message_id2));
+    assert_eq!(result.messages[2].remote_id.as_ref(), Some(&message_id3));
+
+    let conv = Conversation::find_by_id(
+        result.conversation.local_id,
+        &user_ctx.user_stash().connection().await.unwrap(),
+    )
+    .await
+    .unwrap()
+    .unwrap();
+    assert!(
+        conv.labels
+            .iter()
+            .any(|l| l.remote_label_id.as_ref() == Some(&new_label_id))
+    );
 }
 
 // #[test]